+++ src/refactoring/components/admin/coupon/CouponForm.tsx
@@ -1,24 +1,7 @@
-import { useState } from "react"
-import { Coupon } from "../../../../types.ts"
-import { useCouponContext } from "../../../context/CouponContext.tsx"
+import { useCouponForm } from "../../../hooks/useCouponForm.ts"
 
 export default function CouponForm() {
-  const { addCoupon } = useCouponContext()
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: "",
-    code: "",
-    discountType: "percentage",
-    discountValue: 0,
-  })
-  const handleAddCoupon = () => {
-    addCoupon(newCoupon)
-    setNewCoupon({
-      name: "",
-      code: "",
-      discountType: "percentage",
-      discountValue: 0,
-    })
-  }
+  const { newCoupon, handleInputChange, handleAddCoupon } = useCouponForm()
 
   return (
     <div className="space-y-2 mb-4">
@@ -26,20 +9,20 @@ export default function CouponForm() {
         type="text"
         placeholder="쿠폰 이름"
         value={newCoupon.name}
-        onChange={(e) => setNewCoupon({ ...newCoupon, name: e.target.value })}
+        onChange={(e) => handleInputChange("name", e.target.value)}
         className="w-full p-2 border rounded"
       />
       <input
         type="text"
         placeholder="쿠폰 코드"
         value={newCoupon.code}
-        onChange={(e) => setNewCoupon({ ...newCoupon, code: e.target.value })}
+        onChange={(e) => handleInputChange("code", e.target.value)}
         className="w-full p-2 border rounded"
       />
       <div className="flex gap-2">
         <select
           value={newCoupon.discountType}
-          onChange={(e) => setNewCoupon({ ...newCoupon, discountType: e.target.value as "amount" | "percentage" })}
+          onChange={(e) => handleInputChange("discountType", e.target.value as "amount" | "percentage")}
           className="w-full p-2 border rounded"
         >
           <option value="amount">금액(원)</option>
@@ -49,7 +32,7 @@ export default function CouponForm() {
           type="number"
           placeholder="할인 값"
           value={newCoupon.discountValue}
-          onChange={(e) => setNewCoupon({ ...newCoupon, discountValue: parseInt(e.target.value) })}
+          onChange={(e) => handleInputChange("discountValue", parseInt(e.target.value))}
           className="w-full p-2 border rounded"
         />
       </div>
